Remove require command from WC_EBANX_Query_Router and give a namespace
